Revert "Add unit test for extracting services step"
This reverts commit 69663e9e10687554366a2f6ae80bb8d7653a51ad.
