+++ multiapps-controller-process/src/main/java/org/cloudfoundry/multiapps/controller/process/steps/ExtractServicesDeploymentStep.java
@@ -22,17 +22,19 @@ public class ExtractServicesDeploymentStep extends SyncFlowableStep {
     @Override
     protected StepPhase executeStep(ProcessContext context) {
         getStepLogger().debug(Messages.EXTRACT_SERVICES_FROM_BATCH);
-
+        // get singular batch of resources variable
         List<Resource> batchOfResources = context.getVariable(Variables.BATCH_TO_PROCESS);
-
+        // extract all services from the batch of resources
         ServicesCloudModelBuilder servicesCloudModelBuilder = getServicesCloudModelBuilder(context);
         List<CloudServiceInstanceExtended> servicesCalculatedForDeployment = servicesCloudModelBuilder.build(batchOfResources);
+        // Build a list of services for creation and save them in the context:
         List<CloudServiceInstanceExtended> servicesToCreate = servicesCalculatedForDeployment.stream()
                                                                                              .filter(CloudServiceInstanceExtended::isManaged)
                                                                                              .collect(Collectors.toList());
         getStepLogger().debug(Messages.SERVICES_TO_CREATE, SecureSerialization.toJson(servicesToCreate));
         context.setVariable(Variables.SERVICES_TO_CREATE, servicesToCreate);
 
+        // Needed by CreateOrUpdateServicesStep, as it is used as an iteration variable:
         context.setVariable(Variables.SERVICES_TO_CREATE_COUNT, servicesToCreate.size());
 
         return StepPhase.DONE;
