+++ multiapps-controller-process/src/main/java/org/cloudfoundry/multiapps/controller/process/variables/Variables.java
@@ -1,10 +1,13 @@
 package org.cloudfoundry.multiapps.controller.process.variables;
 
 import java.time.Duration;
+import java.util.AbstractMap;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.SortedMap;
 import java.util.UUID;
 
 import org.cloudfoundry.multiapps.controller.client.lib.domain.CloudApplicationExtended;
@@ -590,12 +593,12 @@ public interface Variables {
                                                                                        .defaultValue(true)
                                                                                        .build();
 
-    Variable<List<List<Resource>>> BATCHES_TO_PROCESS = ImmutableJsonStringListVariable.<List<Resource>> builder()
-                                                                                       .name("batchesToProcess")
-                                                                                       .type(new TypeReference<>() {
-                                                                                       })
-                                                                                       .defaultValue(Collections.emptyList())
-                                                                                       .build();
+    ImmutableJsonStringListVariable<List<Resource>> BATCHES_TO_PROCESS = ImmutableJsonStringListVariable.<List<Resource>> builder()
+                                                                                                        .name("batchesToProcess")
+                                                                                                        .type(new TypeReference<>() {
+                                                                                                        })
+                                                                                                        .defaultValue(Collections.emptyList())
+                                                                                                        .build();
     Variable<List<Resource>> BATCH_TO_PROCESS = ImmutableJsonStringVariable.<List<Resource>> builder()
                                                                            .name("batchToProcess")
                                                                            .type(new TypeReference<>() {
