+++ multiapps-controller-process/src/test/java/org/cloudfoundry/multiapps/controller/process/steps/BuildCloudDeployModelStepTest.java
@@ -1,13 +1,24 @@
 package org.cloudfoundry.multiapps.controller.process.steps;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.sap.cloudfoundry.client.facade.domain.CloudServiceKey;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.when;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Stream;
+
 import org.cloudfoundry.multiapps.common.test.TestUtil;
 import org.cloudfoundry.multiapps.common.test.Tester.Expectation;
 import org.cloudfoundry.multiapps.common.util.JsonUtil;
+import org.cloudfoundry.multiapps.controller.client.lib.domain.CloudServiceInstanceExtended;
 import org.cloudfoundry.multiapps.controller.core.cf.util.ModulesCloudModelBuilderContentCalculator;
 import org.cloudfoundry.multiapps.controller.core.cf.v2.ApplicationCloudModelBuilder;
 import org.cloudfoundry.multiapps.controller.core.cf.v2.ServiceKeysCloudModelBuilder;
+import org.cloudfoundry.multiapps.controller.core.cf.v2.ServicesCloudModelBuilder;
 import org.cloudfoundry.multiapps.controller.core.helpers.ModuleToDeployHelper;
 import org.cloudfoundry.multiapps.controller.core.model.DeployedMta;
 import org.cloudfoundry.multiapps.controller.core.test.DescriptorTestUtil;
@@ -19,16 +30,8 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mock;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Stream;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.sap.cloudfoundry.client.facade.domain.CloudServiceKey;
 
 class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeployModelStep> {
 
@@ -39,6 +42,7 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
 
     protected List<Module> modulesToDeploy;
     protected DeployedMta deployedMta;
+    protected List<CloudServiceInstanceExtended> servicesToBind;
     protected Map<String, List<CloudServiceKey>> serviceKeys;
 
     @Mock
@@ -48,13 +52,15 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
     @Mock
     protected ServiceKeysCloudModelBuilder serviceKeysCloudModelBuilder;
     @Mock
+    protected ServicesCloudModelBuilder servicesCloudModelBuilder;
+    @Mock
     protected ModuleToDeployHelper moduleToDeployHelper;
 
     public static Stream<Arguments> testExecute() {
         return Stream.of(
 // @formatter:off
-                Arguments.of(new StepInput("modules-to-deploy-01.json", "services-to-create-01.json", "service-keys-01.json", List.of("api.cf.neo.ondemand.com"), "deployed-mta-12.json")),
-                Arguments.of(new StepInput("modules-to-deploy-01.json", "services-to-create-01.json", "service-keys-01.json", List.of("api.cf.neo.ondemand.com"), null))
+                Arguments.of(new StepInput("modules-to-deploy-01.json", "services-to-bind-01.json", "services-to-create-01.json", "service-keys-01.json", List.of("api.cf.neo.ondemand.com"), "deployed-mta-12.json")),
+                Arguments.of(new StepInput("modules-to-deploy-01.json", "services-to-bind-01.json", "services-to-create-01.json", "service-keys-01.json", List.of("api.cf.neo.ondemand.com"), null))
 // @formatter:on
         );
     }
@@ -68,6 +74,8 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
 
         assertStepFinishedSuccessfully();
 
+        tester.test(() -> context.getVariable(Variables.SERVICES_TO_BIND),
+                    new Expectation(Expectation.Type.JSON, input.servicesToBindLocation));
         tester.test(() -> context.getVariable(Variables.SERVICE_KEYS_TO_CREATE),
                     new Expectation(Expectation.Type.JSON, input.serviceKeysLocation));
         tester.test(() -> context.getVariable(Variables.MODULES_TO_DEPLOY),
@@ -83,6 +91,10 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
         modulesToDeploy = JsonUtil.fromJson(modulesToDeployString, new TypeReference<>() {
         });
 
+        String servicesToBindString = TestUtil.getResourceAsString(input.servicesToBindLocation, getClass());
+        servicesToBind = JsonUtil.fromJson(servicesToBindString, new TypeReference<>() {
+        });
+
         String serviceKeysString = TestUtil.getResourceAsString(input.serviceKeysLocation, getClass());
         serviceKeys = JsonUtil.fromJson(serviceKeysString, new TypeReference<>() {
         });
@@ -94,6 +106,7 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
         when(moduleToDeployHelper.isApplication(any())).thenReturn(true);
         when(modulesCloudModelBuilderContentCalculator.calculateContentForBuilding(any())).thenReturn(modulesToDeploy);
         when(applicationCloudModelBuilder.getApplicationDomains(any(), any())).thenReturn(input.customDomains);
+        when(servicesCloudModelBuilder.build(any())).thenReturn(servicesToBind);
         when(serviceKeysCloudModelBuilder.build()).thenReturn(serviceKeys);
         context.setVariable(Variables.DEPLOYED_MTA, deployedMta);
     }
@@ -113,14 +126,16 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
     protected static class StepInput {
 
         public final String servicesToBindLocation;
+        public final String servicesToCreateLocation;
         public final String deployedMtaLocation;
         public final String serviceKeysLocation;
         public final String modulesToDeployLocation;
         public final List<String> customDomains;
 
-        public StepInput(String modulesToDeployLocation, String servicesToBindLocation, String serviceKeysLocation,
-                         List<String> customDomains, String deployedMtaLocation) {
+        public StepInput(String modulesToDeployLocation, String servicesToBindLocation, String servicesToCreateLocation,
+                         String serviceKeysLocation, List<String> customDomains, String deployedMtaLocation) {
             this.servicesToBindLocation = servicesToBindLocation;
+            this.servicesToCreateLocation = servicesToCreateLocation;
             this.deployedMtaLocation = deployedMtaLocation;
             this.serviceKeysLocation = serviceKeysLocation;
             this.modulesToDeployLocation = modulesToDeployLocation;
@@ -140,6 +155,11 @@ class BuildCloudDeployModelStepTest extends SyncFlowableStepTest<BuildCloudDeplo
                   getModulesContentCalculator(ProcessContext context, List<Module> mtaDescriptorModules, Set<String> mtaManifestModuleNames,
                                               Set<String> deployedModuleNames, Set<String> allMtaModuleNames) {
             return modulesCloudModelBuilderContentCalculator;
+        }
+
+        @Override
+        protected ServicesCloudModelBuilder getServicesCloudModelBuilder(ProcessContext context) {
+            return servicesCloudModelBuilder;
         }
 
         @Override
